Added null check for ApplicationUser AccessKey to TransactionsController actions.

+++ src/API/TradeTracker.Api/Controllers/TransactionsController.cs
@@ -36,78 +36,115 @@ namespace TradeTracker.Api.Controllers
         [HttpPost(Name = "AddTransaction")]
         public async Task<ActionResult<Guid>> AddTransaction([FromBody] CreateTransactionDto createTransactionDto)
         {
+            var accessKey = User.FindFirstValue("AccessKey");
+            if (accessKey == null)
+            {
+                return Unauthorized();
+            }
+
             var command = _mapper.Map<CreateTransactionCommand>(createTransactionDto);
+            command.AccessKey = accessKey;
 
-            command.AccessKey = User.FindFirstValue("AccessKey");
-
-            var transactionToReturn = await _mediator.Send(command);
+            var createdTransaction = await _mediator.Send(command);
             
             return CreatedAtAction(
                 "GetTransactionById",
-                new { id = transactionToReturn.TransactionId },
-                transactionToReturn);
+                new { id = createdTransaction.TransactionId },
+                createdTransaction);
         }
 
         [HttpGet("{id}", Name = "GetTransactionById")]
         public async Task<ActionResult<TransactionForReturnDto>> GetTransactionById(Guid id)
         {
+            var accessKey = User.FindFirstValue("AccessKey");
+            if (accessKey == null)
+            {
+                return Unauthorized();
+            }
+
             var query = new GetTransactionQuery()
             {
-                AccessKey = User.FindFirstValue("AccessKey"),
+                AccessKey = accessKey,
                 TransactionId = id
             };
             
-            return Ok(await _mediator.Send(query));
+            var returnedTransaction = await _mediator.Send(query);
+            return Ok(returnedTransaction);
         }
 
         [HttpGet(Name = "GetTransactionsList")]
         public async Task<ActionResult<PagedTransactionsListVm>> GetTransactionsList(
             [FromQuery] GetTransactionsListResourceParameters getTransactionsListResourceParameters)
         {
+            var accessKey = User.FindFirstValue("AccessKey");
+            if (accessKey == null)
+            {
+                return Unauthorized();
+            }
+
             var query = _mapper.Map<GetTransactionsListQuery>(getTransactionsListResourceParameters);
+            query.AccessKey = accessKey;
 
-            query.AccessKey = User.FindFirstValue("AccessKey");
-
-            return Ok(await _mediator.Send(query));
+            var returnedTransactions = await _mediator.Send(query);
+            return Ok(returnedTransactions);
         }
 
         [HttpPut("{id}", Name = "UpdateTransaction")]
         public async Task<ActionResult> UpdateTransaction(Guid id, [FromBody] UpdateTransactionCommandDto updateTransactionCommandDto)
         {
-            var command = _mapper.Map<UpdateTransactionCommand>(updateTransactionCommandDto);
+            var accessKey = User.FindFirstValue("AccessKey");
+            if (accessKey == null)
+            {
+                return Unauthorized();
+            }
 
-            command.AccessKey = User.FindFirstValue("AccessKey");
+            var command = _mapper.Map<UpdateTransactionCommand>(updateTransactionCommandDto);
+            command.AccessKey = accessKey;
             command.TransactionId = id;
 
             await _mediator.Send(command);
-            
             return NoContent();
         }
 
         [HttpDelete("{id}", Name = "DeleteTransaction")]
         public async Task<ActionResult> DeleteTransaction(Guid id)
         {
+            var accessKey = User.FindFirstValue("AccessKey");
+            if (accessKey == null)
+            {
+                return Unauthorized();
+            }
+
             var command = new DeleteTransactionCommand() 
             {
-                AccessKey = User.FindFirstValue("AccessKey"),
+                AccessKey = accessKey,
                 TransactionId = id
             };
+
             await _mediator.Send(command);
-            
             return NoContent();
         }
 
         [HttpGet("export", Name = "ExportTransactions")]
         [FileResultContentType("text/csv")]
-        public async Task<FileResult> ExportTransactions()
+        public async Task<IActionResult> ExportTransactions()
         {
+            var accessKey = User.FindFirstValue("AccessKey");
+            if (accessKey == null)
+            {
+                return Unauthorized();
+            }
+
             var query = new ExportTransactionsQuery()
             {
-                AccessKey = User.FindFirstValue("AccessKey")
+                AccessKey = accessKey
             };
-            var fileDto = await _mediator.Send(query);
 
-            return File(fileDto.Data, fileDto.ContentType, fileDto.TransactionsExportFileName);
+            var fileExportDto = await _mediator.Send(query);
+            return File(
+                fileExportDto.Data, 
+                fileExportDto.ContentType, 
+                fileExportDto.TransactionsExportFileName);
         }
 
         [HttpOptions(Name = "OptionsTransactions")]
